streamlit_app.py
# streamlit_app.py
import streamlit as st
from PIL import Image
import numpy as np
import cv2 
from scipy.spatial import distance as dist
import time 

# A3用紙の既知の寸法（例: 短辺）をセンチメートルで定義
# 基準オブジェクトの既知の長さとして利用します
KNOWN_WIDTH_CM = 29.7 

# =======================================================
# 📏 【採寸ロジック関数】 A3用紙を基準に計算する
# =======================================================

def measure_clothing(image_np, known_width):
    """
    画像（NumPy配列）を受け取り、A3用紙を基準として服の寸法を計算する関数。
    この関数内の画像処理ロジックを完成させる必要があります。
    """
    
    # ⚠️ A3用紙（基準オブジェクト）検出の複雑なロジックは省略されています ⚠️
    # 実際には、以下の処理が必要です:
    # 1. A3用紙の4隅を検出
    # 2. 検出した4隅を使ってパースペクティブ補正（歪み補正）
    # 3. 補正後の画像から、A3のピクセル幅を測定し、Pixels Per Metricを計算
    # 4. 服の輪郭を検出し、着丈/身幅などの測定点を特定
    
    # 現時点では、デプロイテストのためにダミーの結果を返します
    # -------------------------------------------------------------
    
    # (仮の処理時間)
    time.sleep(2) 
    
    return {
        "着丈": 72.5,
        "身幅": 55.0,
        "肩幅": 48.0,
        "袖丈": 61.2
    }

# =======================================================
# 📱 Streamlit UI 部分
# =======================================================

st.title('👕 服の自動採寸アプリ (A3基準)')
st.subheader('服をA3画用紙に置いて撮影した画像をアップロードしてください。')

# ユーザーからのファイルアップロードを許可
uploaded_file = st.file_uploader("採寸したい服の画像をアップロード", type=['jpg', 'jpeg', 'png'])

if uploaded_file is not None:
    # 画像の表示
    image = Image.open(uploaded_file)
    st.image(image, caption='アップロードされた画像', use_column_width=True)
    
    # PIL ImageをOpenCVが扱えるNumpy配列に変換（BGR形式に変換）
    image_np = np.array(image.convert('RGB')) 
    # OpenCVはBGR形式を使うため、RGBをBGRに変換する処理を開発時に追加する必要があります
    image_np = cv2.cvtColor(image_np, cv2.COLOR_RGB2BGR) 
    
    # 採寸ボタン
    if st.button('採寸開始'):
        # 処理状況を通知
        with st.spinner('計測中...画像解析と計算を行っています。'):
            
            try:
                # 採寸ロジックを呼び出す
                measurements = measure_clothing(image_np, KNOWN_WIDTH_CM)
                
                st.success('採寸が完了しました！')
                
                # 結果を表示
                st.markdown("### 📐 計測結果 (A3基準)")
                for key, value in measurements.items():
                    st.write(f"* **{key}:** {value:.1f} cm")
                    
            except Exception as e:
                # デプロイ後にエラーが出た場合の表示
                st.error(f"計測中にエラーが発生しました。コードを確認してください: {e}")

# 注意書き
st.info('※このアプリは、A3画用紙の既知の寸法を基準としています。')