requirements.txt
streamlit
opencv-python
numpy
scipy
Pillow